requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
PyQt5==5.15.11
PyQt5_sip==12.16.1
pyqtgraph==0.13.7
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
shapely==2.0.6
six==1.17.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.49
ultralytics-thop==2.0.13
urllib3==2.2.3
